refactor: simplify settings pages
